Configuração básica do express e da rota /tasks
