+++ package.json
@@ -11,6 +11,7 @@
   "devDependencies": {
     "@types/cron": "^1.3.0",
     "@types/dockerode": "^2.5.5",
+    "@types/express": "^4.16.0",
     "@types/jest": "^23.3.2",
     "@types/js-yaml": "^3.11.2",
     "@types/logform": "^1.2.0",
@@ -26,6 +27,7 @@
   "dependencies": {
     "cron": "^1.4.1",
     "dockerode": "^2.5.7",
+    "express": "^4.16.3",
     "js-yaml": "^3.12.0",
     "logform": "^1.10.0",
     "node-fetch": "^2.2.0",
